pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "impact_engine_measure"
version = "0.1.0"
description = "A package for evaluating product impact."
authors = [
  { name = "eisenhauerIO", email = "info@eisenhauer.io" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "pandas",
  "pyyaml",
  "statsmodels",
  "pysyncon",
  "causalml",
  "online_retail_simulator @ git+https://github.com/eisenhauerIO/tools-catalog-generator.git",
  "artifact-store @ git+https://github.com/eisenhauerIO/utils-artifact-store.git"
]

[project.urls]
Homepage = "https://github.com/eisenhauerIO/tools-impact-engine"

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "black",
  "ruff",
  "pre-commit"
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["impact_engine_measure"]

[tool.hatch.envs.default]
dependencies = [
  "pytest",
  "ruff",
  "black",
  "pre-commit",
  "sphinx>=7.0.0",
  "sphinx-rtd-theme>=1.3.0",
  "myst-parser>=2.0.0",
  "linkify-it-py>=2.0.0",
  "sphinx-autodoc-typehints>=1.12.0",
  "nbsphinx>=0.9.0",
  "ipykernel>=6.0.0",
  "matplotlib>=3.5.0",
  "seaborn>=0.12.0",
  "jupyter",
  "nbconvert",
]

[tool.hatch.envs.default.scripts]
test = "pytest impact_engine_measure/ -v"
format = "black ."
lint = "ruff check impact_engine_measure/"
docs = "cd documentation && sphinx-build -b html . _build/html"

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

